mod: jwttokeninfo 수정, 아바타 body url 조회 부분 수정

+++ api/src/main/java/com/pfplaybackend/api/avatar/repository/AvatarRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface AvatarRepository extends JpaRepository<Avatar, Long> {
+public interface AvatarRepository extends JpaRepository<Avatar, Integer> {
 
 }
